requirements.txt
python-dotenv
requests
yt-dlp
youtube-transcript-api
